requirements.txt
# Core dependencies for Cognitive PID Framework
openai>=1.3.0
anthropic>=0.7.0
sentence-transformers>=2.2.2
PyYAML>=6.0
gitpython>=3.1.40
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
flake8>=6.1.0
python-dotenv>=1.0.0
tenacity>=8.2.3
numpy>=1.24.0
scikit-learn>=1.3.0
requests>=2.31.0
